+++ src/app/events/page.style.jsx
@@ -3,5 +3,5 @@ import styled from 'styled-components';
 
 export const EventsContainer = styled.div`
   background-image: url('https://res.cloudinary.com/dhv234qct/image/upload/v1728888341/Inno2k24/yupqoznoucyhxwchhbv7.png');
-  ${tw`w-full flex flex-col items-center justify-center bg-cover`}
+  ${tw`w-full flex flex-col items-center justify-center bg-cover pt-36`}
 `;
