+++ src/components/EventsPage/CardComponents/DetailsCard/DetailsCard.style.jsx
@@ -18,8 +18,8 @@ export const DescriptionCardInner = styled.div`
 export const DescriptionCardContent = styled(SmallParagraph)`
   ${tw`w-[812px] max-h-[332px] text-white text-[32px] font-normal font-montserrat [word-wrap: break-word] leading-10 text-left`}
 `;
-
 export const DescriptionCardImage = styled.div`
   background-image: url(${(props) => props.image});
-  ${tw`w-[416px] h-[458px] rounded-[17.82px] shadow flex-col justify-start items-start inline-flex  ml-2`}
+  ${tw`w-[416px] h-[458px] rounded-[17.82px] shadow  flex-col justify-start items-start inline-flex ml-2`}
+  box-shadow: 0 0 10px 2px rgba(255, 255, 255, 0.4);
 `;
