+++ src/config/content/EventsPage/GuestLecture/Data.js
@@ -1,0 +1,101 @@
+export const GL_previewItems = [
+  {
+    id: 1,
+    ImageURL: 'https://picsum.photos/500',
+    PreviewDescription: 'xbvfxvbzfdbzvcbvcbvcbvcbcvb',
+  },
+  {
+    id: 2,
+    ImageURL: 'https://picsum.photos/600',
+    PreviewDescription: 'xbvfxvbzfdbzvcbvcbsvgfsdgdsvcbvcbcvb',
+  },
+  {
+    id: 3,
+    ImageURL: 'https://picsum.photos/400',
+    PreviewDescription: 'xbvfxvbzfdbcvzvczxzvcbvcbvcbvcbcvb',
+  },
+  {
+    id: 4,
+    ImageURL: 'https://picsum.photos/600',
+    PreviewDescription: 'xbvfxvbsvdcszzfdbzvcbvcbvcbvcbcvb',
+  },
+  {
+    id: 5,
+    ImageURL: 'https://picsum.photos/500',
+    PreviewDescription: 'xbvfxvbzfdbzvcbvcbvcbvcbcvb',
+    // other props for Preview
+  },
+  {
+    id: 6,
+    ImageURL: 'https://picsum.photos/600',
+    PreviewDescription: 'xbvfxvbzfdbzvcbvcbsvgfsdgdsvcbvcbcvb',
+  },
+  {
+    id: 7,
+    ImageURL: 'https://picsum.photos/400',
+    PreviewDescription: 'xbvfxvbzfdbcvzvczxzvcbvcbvcbvcbcvb',
+  },
+];
+
+export const GL_descriptionItems = [
+  {
+    id: 1,
+    Heading: 'Event 1',
+    Description: 'Description for Event 1',
+    ImageURL: 'https://picsum.photos/800',
+  },
+  {
+    id: 2,
+    Heading: 'Event 2',
+    Description: 'Description for Event 2',
+    ImageURL: 'https://picsum.photos/900',
+  },
+  {
+    id: 3,
+    Heading: 'Event 3',
+    Description: 'Description for Event 3',
+    ImageURL: 'https://picsum.photos/800',
+  },
+  {
+    id: 4,
+    Heading: 'Event 4',
+    Description: 'Description for Event 4',
+    ImageURL: 'https://picsum.photos/600',
+  },
+  {
+    id: 5,
+    Heading: 'Event 5',
+    Description: 'Description for Event 5',
+    ImageURL: 'https://picsum.photos/800',
+  },
+  {
+    id: 6,
+    Heading: 'Event 6',
+    Description: 'Description for Event 6',
+    ImageURL: 'https://picsum.photos/900',
+  },
+  {
+    id: 7,
+    Heading: 'Event 7',
+    Description: 'Description for Event 7',
+    ImageURL: 'https://picsum.photos/800',
+  },
+  {
+    id: 8,
+    Heading: 'Event 8',
+    Description: 'Description for Event 8',
+    ImageURL: 'https://picsum.photos/600',
+  },
+  {
+    id: 9,
+    Heading: 'Event 9',
+    Description: 'Description for Event 9',
+    ImageURL: 'https://picsum.photos/800',
+  },
+  {
+    id: 10,
+    Heading: 'Event 10',
+    Description: 'Description for Event 10',
+    ImageURL: 'https://picsum.photos/900',
+  },
+];
